add tests

+++ src/main/java/eu/deltasource/internship/hotel/service/RoomService.java
@@ -1,6 +1,8 @@
 package eu.deltasource.internship.hotel.service;
 
 import eu.deltasource.internship.hotel.domain.Room;
+import eu.deltasource.internship.hotel.exception.FailedInitializationException;
+import eu.deltasource.internship.hotel.exception.ItemNotFoundException;
 import eu.deltasource.internship.hotel.repository.RoomRepository;
 
 import java.util.List;
@@ -25,15 +27,24 @@ public class RoomService {
     }
 
     public Room saveRoom(Room room) {
+        if (room == null) {
+            throw new FailedInitializationException("Invalid room !");
+        }
         roomRepository.save(room);
         return roomRepository.findById(room.getRoomId());
     }
 
     public void saveRooms(Room... rooms) {
+        if (rooms == null) {
+            throw new FailedInitializationException("Invalid room !");
+        }
         roomRepository.saveAll(rooms);
     }
 
     public boolean deleteRoom(Room room) {
+        if (room == null) {
+            throw new FailedInitializationException("Invalid room !");
+        }
         return roomRepository.delete(room);
     }
 
@@ -42,6 +53,9 @@ public class RoomService {
     }
 
     public Room updateRoom(Room room) {
+        if (room == null) {
+            throw new FailedInitializationException("Invalid room !");
+        }
         return roomRepository.updateRoom(room);
     }
 }
